REF: misplaced tests (#32713)
